Use a single pvect and index to keep track of key and value
Since we always are storing key/value pairs in the epdb, we can
maintain a single pvect pointer and single index if we know that the
pvects always have an even number of elements.  The only tricky part
is that when items are removed from the epdb, they are added to a free
list maintained as a linked list.  So you have to remove items in the
reverse order as you stored them so that the key index will always be
value index + 1.  The advantage is reduced size of sl nodes.